+++ code/viewmodel/entity/bullet.cpp
@@ -1,17 +1,17 @@
 #include "bullet.h"
 #include "enemy.h"
 
-Bullet::Bullet(int x, int y, const Player* owner, int damage, int speed, Direction initial_dir)
+Bullet::Bullet(float x, float y, const Player* owner, int damage, int speed, Direction initial_dir)
     : Entity(x, y, EntityType::Bullet), owner_(owner), damage_(damage), speed_(speed) {
     switch (initial_dir) {
-        case Direction::Up:    velocity_ = {0.0f, -static_cast<float>(speed_)}; break;
-        case Direction::Down:  velocity_ = {0.0f, static_cast<float>(speed_)};  break;
-        case Direction::Left:  velocity_ = {-static_cast<float>(speed_), 0.0f}; break;
-        case Direction::Right: velocity_ = {static_cast<float>(speed_), 0.0f};  break;
-        case Direction::UpLeft: velocity_ = {-static_cast<float>(speed_ / sqrt(2)), -static_cast<float>(speed_ / sqrt(2))}; break;
-        case Direction::UpRight: velocity_ = {static_cast<float>(speed_ / sqrt(2)), -static_cast<float>(speed_ / sqrt(2))}; break;
-        case Direction::DownLeft: velocity_ = {-static_cast<float>(speed_ / sqrt(2)), static_cast<float>(speed_ / sqrt(2))}; break;
-        case Direction::DownRight: velocity_ = {static_cast<float>(speed_ / sqrt(2)), static_cast<float>(speed_ / sqrt(2))}; break;
+        case Direction::Up:    velocity_ = {0.0f, -(speed_)}; break;
+        case Direction::Down:  velocity_ = {0.0f, (speed_)};  break;
+        case Direction::Left:  velocity_ = {-(speed_), 0.0f}; break;
+        case Direction::Right: velocity_ = {(speed_), 0.0f};  break;
+        case Direction::UpLeft: velocity_ = {-(speed_ / sqrt(2.f)), -(speed_ / sqrt(2.f))}; break;
+        case Direction::UpRight: velocity_ = {(speed_ / sqrt(2.f)), -(speed_ / sqrt(2.f))}; break;
+        case Direction::DownLeft: velocity_ = {-(speed_ / sqrt(2.f)), (speed_ / sqrt(2.f))}; break;
+        case Direction::DownRight: velocity_ = {(speed_ / sqrt(2.f)), (speed_ / sqrt(2.f))}; break;
         default:               velocity_ = {0.0f, 0.0f};        break;
     }
     velocity_ += owner_->getVelocity();
